requirements.txt
# Medical AI Assistant - Dependencies
# Created: 2025-11-19
# Python 3.13 Compatible

# Deep Learning & ML Frameworks
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
transformers>=4.40.0
accelerate>=0.30.0
datasets>=2.19.0
tokenizers>=0.19.0
sentencepiece>=0.2.0
protobuf>=4.25.0

# Medical NLP Models & Tools
scispacy>=0.5.4
biopython>=1.83
medspacy>=1.0.0

# ASR (Automatic Speech Recognition)
openai-whisper>=20231117
librosa>=0.10.2
soundfile>=0.12.1
scipy>=1.13.0
jiwer>=3.0.3

# RAG & Vector Databases
langchain>=0.1.0
langchain-community>=0.0.38
langchain-openai>=0.1.0
faiss-cpu>=1.8.0
chromadb>=0.4.24
sentence-transformers>=2.7.0
rank-bm25>=0.2.2

# LLM Integration
openai>=1.30.0
anthropic>=0.25.0
tiktoken>=0.7.0

# API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1

# MLOps & Experiment Tracking
mlflow>=2.12.0
dvc>=3.50.0
wandb>=0.16.6
tensorboard>=2.16.0

# Monitoring & Metrics
prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=7.0.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.4
pyarrow>=16.0.0

# NLP Tools
spacy>=3.7.4
nltk>=3.8.1
regex>=2024.4.16

# Evaluation Metrics
scikit-learn>=1.4.2
seqeval>=1.2.2
rouge-score>=0.1.2
bert-score>=0.3.13

# Database & Caching
sqlalchemy>=2.0.29
redis>=5.0.4
pymongo>=4.7.0

# Configuration Management
pyyaml>=6.0.1
omegaconf>=2.3.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.6
pytest-cov>=5.0.0
pytest-mock>=3.14.0
httpx>=0.27.0

# Code Quality
black>=24.4.0
flake8>=7.0.0
isort>=5.13.2
mypy>=1.10.0

# Logging
loguru>=0.7.2

# Utilities
tqdm>=4.66.3
click>=8.1.7
rich>=13.7.1

# HTTP & Requests
requests>=2.31.0
aiohttp>=3.9.5

# Security
cryptography>=42.0.5

# Visualization
matplotlib>=3.8.4
seaborn>=0.13.2

# Document Processing
pypdf>=4.2.0
beautifulsoup4>=4.12.3

# Additional ML Tools
optuna>=3.6.1
torchcrf>=1.1.0